Function : Weapon UI (1)
- 무기의 효과를 나타내는 UI 적용 완료
* 무기는 업그레이드만 남음

+++ Assets/3_Scripts/Data/TableData/ItemTable/ItemTableClasses.cs
@@ -40,7 +40,7 @@ namespace ItemTableClassGroup
     [Serializable]
     public class WeaponTableData : ItemTableData
     {
-        public float attackValue;
+        public int attackValue;
         public int additionalEffect;
         public float additionEffectValue;
         public float increaseAttackValue;
